requirements.txt
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.0
PyYAML>=6.0
Jinja2>=3.1
feedparser>=6.0
